Use a random number for the unique label instead.

+++ utilities.h
@@ -23,7 +23,7 @@ inline T DivideCeiling(const T a, const T b)
 template <typename T>
 QString IntegerToZeroPaddedHexQString(const T integer)
 {
-	return QStringLiteral("%1").arg(static_cast<uchar>(integer), sizeof(integer) * 2, 0x10, QLatin1Char('0')).toUpper();
+	return QStringLiteral("%1").arg(integer, sizeof(integer) * 2, 0x10, QLatin1Char('0')).toUpper();
 }
 
 std::array<qreal, 3> QColorTosRGB(const QColor &colour);
